Separate the id and size types
Three roles were all spelled unsigned long or unsigned int, which made
it invisible where a value was a global id, a local index or a count.
Types.h names them: GlobalId is fixed at 64 bit because that is the
width the mesh files use, LocalId at 32 bit, and Size counts entities,
values and bytes. Ranks stay int, which is what MPI uses for them.

The distributor takes ranks as int instead of converting them, the
hard-coded MPI_UNSIGNED_LONG is inferred from GlobalId so the two cannot
drift apart, and numTotalCells() no longer truncates a 64 bit count into
an unsigned int.

The -1 that stood for a missing neighbour becomes InvalidLocalId, so the
face, edge and vertex adjacency no longer carries signed ids for the
sake of one sentinel. Upward, Neighbor and FaceIterator follow.

The test build turns on -Wconversion -Werror, which is what found the
remaining narrowings; the MPI collectives take their int counts through
a checked conversion. -Wsign-conversion stays off: MPI counts ranks in
int while the containers index in size_t, so it fires on every rank loop
without pointing at a defect.

AI-generated. Model: Opus 5

+++ Hdf5Reader.h
@@ -263,7 +263,7 @@ class Hdf5Reader {
     hid_t h5space = H5Dget_space(h5dataset);
     checkH5Err(h5space);
     const auto dimcount = H5Sget_simple_extent_ndims(h5space);
-    if (dimcount != 1 + sizes.size()) {
+    if (dimcount != static_cast<int>(1 + sizes.size())) {
       logError() << "Dataset must have" << 1 + sizes.size() << "dimension(s), but it has"
                  << dimcount;
     }
@@ -298,7 +298,7 @@ class Hdf5Reader {
     checkH5Err(
         H5Sselect_hyperslab(h5space, H5S_SELECT_SET, start.data(), nullptr, count.data(), nullptr));
 
-    hid_t h5memspace = H5Screate_simple(count.size(), count.data(), nullptr);
+    hid_t h5memspace = H5Screate_simple(static_cast<int>(count.size()), count.data(), nullptr);
     checkH5Err(h5memspace);
 
     hid_t h5alist = H5Pcreate(H5P_DATASET_XFER);
